# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: EXPP API
  version: "1.0.0"
  description: >
    EXPP — Educational Workflow Platform v1.
    All endpoints require authentication via the `expp_session` HTTP-only cookie
    unless noted as public.
  contact:
    name: EXPP Project
servers:
  - url: /api
    description: Next.js App Router API
...
components:
  schemas:
    ApiSuccess:
      type: object
      required: [success, data, traceId]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          description: Response payload (schema varies per endpoint)
        traceId:
          type: string
    ApiError:
      type: object
      required: [success, error, code, traceId]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
        code:
          type: string
          enum: [AUTH_ERROR, FORBIDDEN, NOT_FOUND, VALIDATION_ERROR, RATE_LIMITED]
        traceId:
          type: string
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: expp_session
security:
  - sessionCookie: []

# ── Paths ──────────────────────────────────────────────────────────────────
paths:
...
  /materials:
    get:
      tags: [materials]
      summary: List teacher's materials
      description: TEACHER role required. Returns only materials owned by the authenticated teacher.
      responses:
        "200":
          description: Material list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }
        "403":
          description: Not a teacher
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }

    post:
      tags: [materials]
      summary: Upload a material (PDF, Word, Excel, PowerPoint, TXT, MD, CSV)
      description: TEACHER role required. Max 50 MB for binary formats and 10 MB for text formats.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
      responses:
        "201":
          description: Material uploaded and text extracted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }
        "422":
          description: Unsupported file type or size exceeded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /materials/{id}:
    get:
      tags: [materials]
      summary: Get material detail including extracted text
      description: TEACHER role required. Returns ForbiddenError if not the owner.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Material detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }
        "403":
          description: Not the owner
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }

    delete:
      tags: [materials]
      summary: Delete a material
      description: TEACHER role required. Owner only.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }
        "403":
          description: Not the owner
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }

  # ── Generation ────────────────────────────────────────────────────────────
  /generation:
    get:
      tags: [generation]
      summary: List generation requests for the authenticated teacher
      responses:
        "200":
          description: Generation request list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }

    post:
      tags: [generation]
      summary: Create a new generation request
      description: TEACHER role required. Triggers background generation pipeline.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [constraints, materialIds]
              properties:
                constraints:
                  type: object
                  description: GenerationConstraints (topic, difficulty, format, questionCount, etc.)
                materialIds:
                  type: array
                  items: { type: string }
      responses:
        "201":
          description: Generation request created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }
        "422":
          description: Invalid constraints
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /generation/{id}:
    get:
      tags: [generation]
      summary: Get generation request detail (plan + result)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Generation detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }
        "403":
          description: Not the owner
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...